+++ compiler/src/iree/compiler/Codegen/Common/CPU/CMakeLists.txt
@@ -82,7 +82,6 @@ iree_cc_library(
     iree::compiler::Codegen::Utils
     iree::compiler::Dialect::HAL::IR
     iree::compiler::Dialect::LinalgExt::IR
-    iree::compiler::Dialect::LinalgExt::Transforms
     iree::compiler::Dialect::LinalgExt::Utils
   PUBLIC
 )
