+++ compiler/src/iree/compiler/Codegen/Common/GPU/CMakeLists.txt
@@ -113,8 +113,6 @@ iree_cc_library(
     iree::compiler::Codegen::Utils
     iree::compiler::Codegen::Utils::VectorOpUtils
     iree::compiler::Dialect::HAL::IR
-    iree::compiler::Dialect::LinalgExt::IR
-    iree::compiler::Dialect::LinalgExt::Transforms
   PUBLIC
 )
