+++ compiler/src/iree/compiler/Codegen/LLVMGPU/LLVMGPUTensorCoreVectorization.cpp
@@ -12,7 +12,6 @@
 #include "iree/compiler/Codegen/Utils/MarkerUtils.h"
 #include "iree/compiler/Codegen/Utils/Utils.h"
 #include "iree/compiler/Dialect/LinalgExt/Transforms/Passes.h"
-#include "iree/compiler/Dialect/LinalgExt/Transforms/Transforms.h"
 #include "llvm/Support/Debug.h"
 #include "mlir/Conversion/VectorToGPU/VectorToGPU.h"
 #include "mlir/Dialect/Linalg/Transforms/Transforms.h"
@@ -34,8 +33,7 @@ namespace mlir::iree_compiler {
 static void vectorizeLinalgOps(mlir::FunctionOpInterface funcOp) {
   MLIRContext *context = funcOp.getContext();
   IRRewriter rewriter(context);
-  IREE::LinalgExt::LinalgTransformationFilter f(
-      StringAttr::get(context, getVectorizeMarker()));
+  LinalgTransformationFilter f(StringAttr::get(context, getVectorizeMarker()));
 
   funcOp.walk([&](Operation *op) {
     if (failed(f.checkAndNotify(rewriter, op)) ||
