+++ compiler/src/iree/compiler/Codegen/Transforms/CMakeLists.txt
@@ -41,7 +41,6 @@ iree_cc_library(
     iree::compiler::Codegen::Utils
     iree::compiler::Dialect::Flow::IR
     iree::compiler::Dialect::HAL::IR
-    iree::compiler::Dialect::LinalgExt::Transforms
   PUBLIC
 )
