+++ compiler/src/iree/compiler/Codegen/Utils/MarkerUtils.cpp
@@ -6,12 +6,87 @@
 
 #include "iree/compiler/Codegen/Utils/MarkerUtils.h"
 
-#include "iree/compiler/Dialect/LinalgExt/Transforms/Passes.h"
 #include "mlir/Dialect/Linalg/Transforms/Transforms.h"
 #include "mlir/IR/Attributes.h"
 #include "mlir/IR/Operation.h"
 
 namespace mlir::iree_compiler {
+
+// Marker used as attribute name in generated Linalg rewriting transformations.
+const StringLiteral LinalgTransforms::kLinalgTransformMarker =
+    "__internal_linalg_transform__";
+
+LinalgTransformationFilter::LinalgTransformationFilter(
+    ArrayRef<StringAttr> matchDisjunction,
+    std::optional<StringAttr> replacement)
+    : matchDisjunction(matchDisjunction.begin(), matchDisjunction.end()),
+      replacement(replacement), matchByDefault(false) {}
+
+LinalgTransformationFilter::LinalgTransformationFilter(
+    const FilterFunction &f, ArrayRef<StringAttr> matchDisjunction,
+    std::optional<StringAttr> replacement)
+    : matchDisjunction(matchDisjunction.begin(), matchDisjunction.end()),
+      replacement(replacement), matchByDefault(false) {
+  if (f) {
+    filters.push_back(f);
+  }
+}
+
+LogicalResult LinalgTransformationFilter::checkAndNotify(RewriterBase &rewriter,
+                                                         Operation *op) const {
+  if (llvm::any_of(filters,
+                   [&](const FilterFunction &f) { return failed(f(op)); })) {
+    return failure();
+  }
+
+  auto attr = op->template getAttrOfType<StringAttr>(
+      LinalgTransforms::kLinalgTransformMarker);
+
+  if (!attr) {
+    // 1. Has no filter case and matchDisjunction is empty.
+    if (matchDisjunction.empty() || matchByDefault) {
+      return success();
+    }
+
+    // 2. Has no filter but was expecting a filter.
+    return rewriter.notifyMatchFailure(op, [&](Diagnostic &diag) {
+      diag << " does not have any filter from list: ";
+      interleaveComma(matchDisjunction, diag);
+    });
+  }
+
+  // 4. Match explicit filter.
+  for (auto filter : matchDisjunction) {
+    if (attr.getValue() == filter) {
+      return success();
+    }
+  }
+
+  // 5. Fail to match.
+  return rewriter.notifyMatchFailure(op, [&](Diagnostic &diag) {
+    diag << " does not have any filter from list: ";
+    interleaveComma(matchDisjunction, diag);
+  });
+}
+
+void LinalgTransformationFilter::replaceLinalgTransformationFilter(
+    RewriterBase &rewriter, Operation *op) const {
+  if (replacement.has_value()) {
+    op->setAttr(LinalgTransforms::kLinalgTransformMarker, replacement.value());
+  } else {
+    op->removeAttr(
+        rewriter.getStringAttr(LinalgTransforms::kLinalgTransformMarker));
+  }
+}
+
+bool LinalgTransformationFilter::hasReplacementFilter(Operation *op) const {
+  if (!replacement) {
+    return false;
+  }
+  auto attr = op->getAttr(LinalgTransforms::kLinalgTransformMarker)
+                  .dyn_cast<StringAttr>();
+  return attr && attr == *replacement;
+}
 
 struct VectorTransforms {
   static const StringLiteral kVectorTransformMarker;
@@ -46,16 +121,16 @@ StringRef getVectorizeMarker() { return "vectorize"; }
 StringRef getDeleteMarker() { return "delete"; }
 
 StringRef getMarkerOrNull(Operation *op) {
-  StringAttr attr = op->getAttrOfType<StringAttr>(
-      IREE::LinalgExt::LinalgTransforms::kLinalgTransformMarker);
+  StringAttr attr =
+      op->getAttrOfType<StringAttr>(LinalgTransforms::kLinalgTransformMarker);
   if (!attr)
     return "";
   return attr.getValue();
 }
 
 bool hasMarker(Operation *op, ArrayRef<StringRef> marker) {
-  StringAttr attr = op->getAttrOfType<StringAttr>(
-      IREE::LinalgExt::LinalgTransforms::kLinalgTransformMarker);
+  StringAttr attr =
+      op->getAttrOfType<StringAttr>(LinalgTransforms::kLinalgTransformMarker);
   return attr && (marker.empty() ||
                   llvm::any_of(marker, [&attr](StringRef markerValue) {
                     return attr.getValue() == markerValue;
@@ -63,7 +138,7 @@ bool hasMarker(Operation *op, ArrayRef<StringRef> marker) {
 }
 
 void setMarker(Operation *op, StringRef marker) {
-  op->setAttr(IREE::LinalgExt::LinalgTransforms::kLinalgTransformMarker,
+  op->setAttr(LinalgTransforms::kLinalgTransformMarker,
               StringAttr::get(op->getContext(), marker));
 }
